first pass at c2enc and c2dec, all builds OK but now need to unit test quantise.c and codec2.c functions

+++ src/codec2.h
@@ -36,6 +36,6 @@
 void *codec2_create();
 void codec2_destroy(void *codec2_state);
 void codec2_encode(void *codec2_state, char bits[], short speech_in[]);
-void codec2_decode((void *codec2_state, int speech_out[], char bits[]);
+void codec2_decode(void *codec2_state, short speech_out[], char bits[]);
 
 #endif
